Completed first draft
This is the basic template needed to start implementing functionality in a future branch. There are lots of improvements yet to be made, like finding ways not to repeat content on the screen. I need to get more inspiration on what menu pages look like for that. Later though. Merging into branch 3