Initialised Database Schema
Created a database on my local host and connected the project to it. The issue is that spring refuses to create the database for us for some reason, so it needs to be created manually. The database will be called team7_soho_kids_database. You can open MariaDB MySQL client and type CREATE SCHEMA IF NOT EXISTS team7_soho_kids_database; . Spring boot should be able to connect to that from now on. Of course, this means we'll all have different databases with different data, so weshould reset the tables (during development only) and have the program populate them again everytime. I'll do that in the schema.sql file in the next commit. When you use tables, just update the schema.sql file to create the tables you need. I've done mine.