Added untested link controller
Two methods added - first is the controller which links the two pages by accepting button inputs to a new url - home/submissions/id which hosts the details template. The other checks whether the light image counter part of the drawing image exists. If so, the controller uses this to also add the light image to the view.