Comment on previous commit
I did this using filter classes that reset the flex properties of elements they are applied to through JS. The reason images were appearing off-centred is because the reset classes weren't being applied to the items being flexed, but rather the children of those items. changed the DOM so that not as many descendents were used, which made it simpler to access the elements that were actually flexed. The other problem still remains though. You can't deselect a filter or stack multiple filters. That should be easy enough to implement if needed. --- Also made the filter buttons glow gold when active