Async Comment posting and retrieval successful
Now comments are posted & retrieved without reloading the page. In this commit, I finished off the move from my local main (which was developed on accidently - time pressure stress I suppose) and wired the JS file to the submission-details html. The main bit of code in this one was the implementation of the fetchComments JS function