Finalised the Async Rating system
In summary, the like function was completely detached from the rating function, which allows individual async requests to be made separately. RatingClass & it's uses were changed to omit the likes field, and likeCounts are now stored in the database far more efficiently as actual counts. Glad I got this done