Began Documentation & Reinstating old onpageload comment fetching
Keeping the AJAX of course, but for some reason onload attributes & event listeners aren't working when the page is loaded / refreshed, so I've chosen to keep the old method of displaying them when thios happens, but implementing AJAX on top whenever there's a user interaction beyond a refresh. The two systems work together nicely.