Finally figured out why some tests where failing when they shouldn't
When I don't drop the tables, I think the primary key autoincrement info is retained somewhere. This meant that once entried where delted, their ids would know longer be repopulate, and instead increment to another value. This meant that selecting 1L as an id and assuming it to be there was a mistake. The assumption was wrong. Instead, I changed the SQL queries to not use hardcoded values, but rather obtain a value that is in the database already and use that for the id.